requirements.txt
typer[all]
rich
google-genai
openai
python-dotenv
